+++ bindings/matrix-sdk-ffi/src/room.rs
@@ -139,6 +139,13 @@ impl Room {
         }
     }
 
+    /// Forces the current outbound group session to be discarded such
+    /// that another one will be created next time an event is sent.
+    pub async fn force_discard_session(&self) -> Result<(), ClientError> {
+        self.inner.discard_room_key().await?;
+        Ok(())
+    }
+
     pub async fn timeline(&self) -> Arc<Timeline> {
         let mut write_guard = self.timeline.write().await;
         if let Some(timeline) = &*write_guard {